feat: enable both system modules for imageOutdated and updateCheck + move drivers to their own modules
