+++ drv/flatpak/flatpak.go
@@ -1,14 +1,16 @@
-package drv
+package flatpak
 
 import (
 	"log/slog"
+	"os/exec"
 	"strings"
 
+	. "github.com/ublue-os/uupd/drv/generic"
 	"github.com/ublue-os/uupd/pkg/percent"
 	"github.com/ublue-os/uupd/pkg/session"
 )
 
-type DistroboxUpdater struct {
+type FlatpakUpdater struct {
 	Config       DriverConfiguration
 	Tracker      *TrackerConfiguration
 	binaryPath   string
@@ -16,7 +18,7 @@ type DistroboxUpdater struct {
 	usersEnabled bool
 }
 
-func (up DistroboxUpdater) Steps() int {
+func (up FlatpakUpdater) Steps() int {
 	if up.Config.Enabled {
 		var steps = 1
 		if up.usersEnabled {
@@ -27,24 +29,24 @@ func (up DistroboxUpdater) Steps() int {
 	return 0
 }
 
-func (up DistroboxUpdater) New(config UpdaterInitConfiguration) (DistroboxUpdater, error) {
-	userdesc := "Distroboxes for User:"
+func (up FlatpakUpdater) New(config UpdaterInitConfiguration) (FlatpakUpdater, error) {
+	userdesc := "Apps for User:"
 	up.Config = DriverConfiguration{
-		Title:           "Distrobox",
-		Description:     "Rootful Distroboxes",
+		Title:           "Flatpak",
+		Description:     "System Apps",
 		UserDescription: &userdesc,
 		Enabled:         true,
 		MultiUser:       true,
 		DryRun:          config.DryRun,
 		Environment:     config.Environment,
 	}
-	up.Config.logger = config.Logger.With(slog.String("module", strings.ToLower(up.Config.Title)))
+	up.Config.Logger = config.Logger.With(slog.String("module", strings.ToLower(up.Config.Title)))
 	up.usersEnabled = false
 	up.Tracker = nil
 
-	binaryPath, exists := up.Config.Environment["UUPD_DISTROBOX_BINARY"]
+	binaryPath, exists := up.Config.Environment["UUPD_FLATPAK_BINARY"]
 	if !exists || binaryPath == "" {
-		up.binaryPath = "/usr/bin/distrobox"
+		up.binaryPath = "/usr/bin/flatpak"
 	} else {
 		up.binaryPath = binaryPath
 	}
@@ -52,16 +54,16 @@ func (up DistroboxUpdater) New(config UpdaterInitConfiguration) (DistroboxUpdate
 	return up, nil
 }
 
-func (up *DistroboxUpdater) SetUsers(users []session.User) {
+func (up *FlatpakUpdater) SetUsers(users []session.User) {
 	up.users = users
 	up.usersEnabled = true
 }
 
-func (up DistroboxUpdater) Check() (bool, error) {
+func (up FlatpakUpdater) Check() (bool, error) {
 	return true, nil
 }
 
-func (up DistroboxUpdater) Update() (*[]CommandOutput, error) {
+func (up FlatpakUpdater) Update() (*[]CommandOutput, error) {
 	var finalOutput = []CommandOutput{}
 
 	if up.Config.DryRun {
@@ -77,8 +79,9 @@ func (up DistroboxUpdater) Update() (*[]CommandOutput, error) {
 	}
 
 	percent.ChangeTrackerMessageFancy(*up.Tracker.Writer, up.Tracker.Tracker, up.Tracker.Progress, percent.TrackerMessage{Title: up.Config.Title, Description: up.Config.Description})
-	cli := []string{up.binaryPath, "upgrade", "-a"}
-	out, err := session.RunUID(up.Config.logger, slog.LevelDebug, 0, cli, nil)
+	cli := []string{up.binaryPath, "update", "-y", "--noninteractive"}
+	flatpakCmd := exec.Command(cli[0], cli[1:]...)
+	out, err := session.RunLog(up.Config.Logger, slog.LevelDebug, flatpakCmd)
 	tmpout := CommandOutput{}.New(out, err)
 	tmpout.Context = up.Config.Description
 	tmpout.Cli = cli
@@ -89,9 +92,9 @@ func (up DistroboxUpdater) Update() (*[]CommandOutput, error) {
 	for _, user := range up.users {
 		up.Tracker.Tracker.IncrementSection(err)
 		context := *up.Config.UserDescription + " " + user.Name
-		percent.ChangeTrackerMessageFancy(*up.Tracker.Writer, up.Tracker.Tracker, up.Tracker.Progress, percent.TrackerMessage{Title: up.Config.Title, Description: *up.Config.UserDescription + " " + user.Name})
-		cli := []string{up.binaryPath, "upgrade", "-a"}
-		out, err := session.RunUID(up.Config.logger, slog.LevelDebug, user.UID, cli, nil)
+		percent.ChangeTrackerMessageFancy(*up.Tracker.Writer, up.Tracker.Tracker, up.Tracker.Progress, percent.TrackerMessage{Title: up.Config.Title, Description: context})
+		cli := []string{up.binaryPath, "update", "-y"}
+		out, err := session.RunUID(up.Config.Logger, slog.LevelDebug, user.UID, cli, nil)
 		tmpout = CommandOutput{}.New(out, err)
 		tmpout.Context = context
 		tmpout.Cli = cli
@@ -99,12 +102,4 @@ func (up DistroboxUpdater) Update() (*[]CommandOutput, error) {
 		finalOutput = append(finalOutput, *tmpout)
 	}
 	return &finalOutput, nil
-}
-
-func (up *DistroboxUpdater) Logger() *slog.Logger {
-	return up.Config.logger
-}
-
-func (up *DistroboxUpdater) SetLogger(logger *slog.Logger) {
-	up.Config.logger = logger
 }
