+++ drv/rpmostree/rpmostree.go
@@ -1,15 +1,17 @@
-package drv
+package rpmostree
 
 // Temporary: WILL get removed at some point.
 // FIXME: Remove this on Spring 2025 when we all move to dnf5 and bootc ideally
 
 import (
 	"encoding/json"
-	"github.com/ublue-os/uupd/pkg/session"
 	"log/slog"
 	"os/exec"
 	"strings"
 	"time"
+
+	. "github.com/ublue-os/uupd/drv/generic"
+	"github.com/ublue-os/uupd/pkg/session"
 )
 
 type rpmOstreeStatus struct {
@@ -31,7 +33,7 @@ func (up RpmOstreeUpdater) Outdated() (bool, error) {
 	var timestamp time.Time
 
 	cmd := exec.Command(up.BinaryPath, "status", "--json", "--booted")
-	out, err := session.RunLog(up.Config.logger, slog.LevelDebug, cmd)
+	out, err := session.RunLog(up.Config.Logger, slog.LevelDebug, cmd)
 	if err != nil {
 		return false, err
 	}
@@ -50,7 +52,7 @@ func (up RpmOstreeUpdater) Update() (*[]CommandOutput, error) {
 	var cmd *exec.Cmd
 	binaryPath := up.BinaryPath
 	cli := []string{binaryPath, "upgrade"}
-	out, err := session.RunLog(up.Config.logger, slog.LevelDebug, cmd)
+	out, err := session.RunLog(up.Config.Logger, slog.LevelDebug, cmd)
 	tmpout := CommandOutput{}.New(out, err)
 	tmpout.Cli = cli
 	tmpout.Failure = err != nil
@@ -66,20 +68,6 @@ func (up RpmOstreeUpdater) Steps() int {
 	return 0
 }
 
-func BootcCompatible(binaryPath string) (bool, error) {
-	cmd := exec.Command(binaryPath, "status", "--format=json")
-	out, err := cmd.CombinedOutput()
-	if err != nil {
-		return false, nil
-	}
-	var status bootcStatus
-	err = json.Unmarshal(out, &status)
-	if err != nil {
-		return false, nil
-	}
-	return !(status.Status.Booted.Incompatible || status.Status.Staged.Incompatible), nil
-}
-
 func (up RpmOstreeUpdater) New(config UpdaterInitConfiguration) (RpmOstreeUpdater, error) {
 	up.Config = DriverConfiguration{
 		Title:       "System",
@@ -88,7 +76,7 @@ func (up RpmOstreeUpdater) New(config UpdaterInitConfiguration) (RpmOstreeUpdate
 		DryRun:      config.DryRun,
 		Environment: config.Environment,
 	}
-	up.Config.logger = config.Logger.With(slog.String("module", strings.ToLower(up.Config.Title)))
+	up.Config.Logger = config.Logger.With(slog.String("module", strings.ToLower(up.Config.Title)))
 	if up.Config.DryRun {
 		return up, nil
 	}
@@ -117,14 +105,6 @@ func (up RpmOstreeUpdater) Check() (bool, error) {
 	}
 
 	updateNecessary := strings.Contains(string(out), "AvailableUpdate")
-	up.Config.logger.Debug("Executed update check", slog.String("output", string(out)), slog.Bool("update", updateNecessary))
+	up.Config.Logger.Debug("Executed update check", slog.String("output", string(out)), slog.Bool("update", updateNecessary))
 	return updateNecessary, nil
-}
-
-func (up *RpmOstreeUpdater) Logger() *slog.Logger {
-	return up.Config.logger
-}
-
-func (up *RpmOstreeUpdater) SetLogger(logger *slog.Logger) {
-	up.Config.logger = logger
 }
