+++ drv/system/system.go
@@ -1,4 +1,4 @@
-package drv
+package system
 
 import (
 	"encoding/json"
@@ -7,6 +7,8 @@ import (
 	"strings"
 	"time"
 
+	. "github.com/ublue-os/uupd/drv/generic"
+	"github.com/ublue-os/uupd/drv/rpmostree"
 	"github.com/ublue-os/uupd/pkg/session"
 )
 
@@ -34,8 +36,6 @@ type SystemUpdateDriver interface {
 	Outdated() (bool, error)
 	Check() (bool, error)
 	Update() (*[]CommandOutput, error)
-	Logger() *slog.Logger
-	SetLogger(value *slog.Logger)
 }
 
 type SystemUpdater struct {
@@ -77,9 +77,9 @@ func (up SystemUpdater) Update() (*[]CommandOutput, error) {
 	var cmd *exec.Cmd
 	binaryPath := up.BinaryPath
 	cli := []string{binaryPath, "upgrade", "--quiet"}
-	up.Config.logger.Debug("Executing update", slog.Any("cli", cli))
+	up.Config.Logger.Debug("Executing update", slog.Any("cli", cli))
 	cmd = exec.Command(cli[0], cli[1:]...)
-	out, err := session.RunLog(up.Config.logger, slog.LevelDebug, cmd)
+	out, err := session.RunLog(up.Config.Logger, slog.LevelDebug, cmd)
 	tmpout := CommandOutput{}.New(out, err)
 	tmpout.Failure = err != nil
 	tmpout.Context = "System Update"
@@ -94,14 +94,6 @@ func (up SystemUpdater) Steps() int {
 	return 0
 }
 
-func EnvOrFallback(environment EnvironmentMap, key string, fallback string) string {
-	validCase, exists := environment[key]
-	if exists && validCase != "" {
-		return validCase
-	}
-	return fallback
-}
-
 func (up SystemUpdater) New(config UpdaterInitConfiguration) (SystemUpdater, error) {
 	up.Config = DriverConfiguration{
 		Title:       "System",
@@ -110,7 +102,7 @@ func (up SystemUpdater) New(config UpdaterInitConfiguration) (SystemUpdater, err
 		DryRun:      config.DryRun,
 		Environment: config.Environment,
 	}
-	up.Config.logger = config.Logger.With(slog.String("module", strings.ToLower(up.Config.Title)))
+	up.Config.Logger = config.Logger.With(slog.String("module", strings.ToLower(up.Config.Title)))
 	up.BinaryPath = EnvOrFallback(config.Environment, "UUPD_BOOTC_BINARY", "/usr/bin/bootc")
 
 	return up, nil
@@ -128,14 +120,59 @@ func (up SystemUpdater) Check() (bool, error) {
 	}
 
 	updateNecessary := !strings.Contains(string(out), "No changes in:")
-	up.Config.logger.Debug("Executed update check", slog.String("output", string(out)), slog.Bool("update", updateNecessary))
+	up.Config.Logger.Debug("Executed update check", slog.String("output", string(out)), slog.Bool("update", updateNecessary))
 	return updateNecessary, nil
 }
 
-func (up *SystemUpdater) Logger() *slog.Logger {
-	return up.Config.logger
+func BootcCompatible(binaryPath string) (bool, error) {
+	cmd := exec.Command(binaryPath, "status", "--format=json")
+	out, err := cmd.CombinedOutput()
+	if err != nil {
+		return false, nil
+	}
+	var status bootcStatus
+	err = json.Unmarshal(out, &status)
+	if err != nil {
+		return false, nil
+	}
+	return !(status.Status.Booted.Incompatible || status.Status.Staged.Incompatible), nil
 }
 
-func (up *SystemUpdater) SetLogger(logger *slog.Logger) {
-	up.Config.logger = logger
+func InitializeSystemDriver(initConfiguration UpdaterInitConfiguration) (SystemUpdateDriver, DriverConfiguration, bool, error) {
+	var enableUpd bool = true
+
+	rpmOstreeUpdater, err := rpmostree.RpmOstreeUpdater{}.New(initConfiguration)
+	if err != nil {
+		enableUpd = false
+	}
+
+	systemUpdater, err := SystemUpdater{}.New(initConfiguration)
+	if err != nil {
+		enableUpd = false
+	}
+
+	isBootc, err := BootcCompatible(systemUpdater.BinaryPath)
+	if err != nil {
+		isBootc = false
+	}
+
+	if !isBootc {
+		slog.Debug("Using rpm-ostree fallback as system driver")
+	}
+
+	// The system driver to be applied needs to have the correct "enabled" value since it will NOT update from here onwards.
+	systemUpdater.Config.Enabled = systemUpdater.Config.Enabled && isBootc && enableUpd
+	rpmOstreeUpdater.Config.Enabled = rpmOstreeUpdater.Config.Enabled && !isBootc && enableUpd
+
+	var finalConfig DriverConfiguration
+	var mainSystemDriver SystemUpdateDriver
+	if isBootc {
+		mainSystemDriver = &systemUpdater
+		finalConfig = systemUpdater.Config
+	} else {
+		mainSystemDriver = &rpmOstreeUpdater
+		finalConfig = systemUpdater.Config
+	}
+
+	return mainSystemDriver, finalConfig, isBootc, err
 }
