+++ uupd.spec
@@ -39,8 +39,7 @@ install -Dpm 644 %{name}.timer %{buildroot}%{_unitdir}/%{name}.timer
 install -Dpm 644 %{name}.rules %{buildroot}%{_sysconfdir}/polkit-1/rules.d/%{name}.rules
 
 %check
-# go test should be here if you have tests, e.g.
-# go test -v ./...
+go test -v ./...
 
 %post
 %systemd_post %{name}.timer
